pyproject.toml
[project]
name = "pms-backend"
version = "0.1.0"
description = "Performance Management System — KPI Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic[email]>=2.7.0",
    "pydantic-settings>=2.2.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "redis[asyncio]>=5.0.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "apscheduler>=3.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "aiosqlite>=0.20.0",
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

